Change id from Int to Long

+++ app/asiproduct/ASIProduct.scala
@@ -2,6 +2,6 @@ package asiproduct
 
 import play.api.libs.json.JsValue
 
-case class ASIProduct(internalId: Option[Int] = None, rawData: JsValue, Id: Long, Name: String, Description: String)
+case class ASIProduct(internalId: Option[Long] = None, rawData: JsValue, Id: Long, Name: String, Description: String)
 
-case class ASIQuoteProduct(id: Option[Int] = None, quoteId: Int, productInternalId: Int)
+case class ASIQuoteProduct(id: Option[Long] = None, quoteId: Long, productInternalId: Long)
