+++ app/user/Role.scala
@@ -1,6 +1,6 @@
 package user
 
 
-case class Role(role: String, userId: Int) {
+case class Role(role: String, userId: Long) {
 
 }
